Narrow hex/base64/binary format validation to RC4/RC4Drop only
Move validateFormatInput() out of Utils.convertToByteArray()/convertToByteString(),
which are shared by ~50 operations, and into Ciphers.mjs where it's only
reachable via parseFormatString() (used by RC4/RC4Drop). This keeps the fix
scoped to issue #2004 instead of changing accepted input globally, and avoids
the unrelated HMAC/Scrypt regressions the broader validation caused.
